third_party/WebKit/LayoutTests/virtual/feature-policy/http/tests/feature-policy/vibrate_in_cross_origin_iframe_blocked-expected.txt

CONSOLE ERROR: line 10: Navigator.vibrate() is not enabled in feature policy for this frame.
